refactor: portfolio dto & vo 변수 타입 변경

+++ src/main/java/com/be/portfolio/dto/req/PortfolioReqDto.java
@@ -19,7 +19,7 @@ public class PortfolioReqDto {
     private String portfolioName;
     private double total;
     private double expectedReturn;
-    private double riskLevel;
+    private int riskLevel;
     private long memberNum;
 
     private List<PortfolioItemReqDto> portfolioItems;
